--- src/main/mainRpcClient/getRpcConfigurationsFromDisk/getBitcoinConf.ts
@@ -1,22 +1,0 @@
-import { promises as fs } from "fs";
-import { RPC_ERROR } from "_c/constants";
-import { ErrorWithCode } from "_c/ErrorWithCode";
-
-export const getBitcoinConf = async (dataDir: string) => {
-  try {
-    return await fs.readFile(`${dataDir}/bitcoin.conf`, {
-      encoding: "utf8",
-    });
-  } catch (error) {
-    /* istanbul ignore else */
-    if (error.code === "ENOENT") {
-      throw new ErrorWithCode(
-        "Could not open `bitcoin.conf`",
-        RPC_ERROR.couldNotOpenBitcoinConf,
-      );
-    }
-
-    /* istanbul ignore next */
-    throw error;
-  }
-};
