--- src/main/mainRpcClient/getRpcConfigurationsFromDisk/getCookiePath.ts
@@ -1,13 +1,0 @@
-export const getCookiePath = (chain: string, dataDir: string) => {
-  let networkDir: string;
-
-  /* istanbul ignore if */ if (chain === "testnet") {
-    networkDir = `${dataDir}/testnet3/`;
-  } /* istanbul ignore if */ else if (chain === "regtest") {
-    networkDir = `${dataDir}/regtest/`;
-  } else {
-    networkDir = `${dataDir}/`;
-  }
-
-  return `${networkDir}.cookie`;
-};
